+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Un-break workflows that use `format = "file_fast"` (#1339, @koefoeden).
 * Fix deadlock in `error = "trim"` (#1340, @koefoeden).
+* Remove tailored debugging message (#1341, @koefoeden).
 
 # targets 1.8.0